requirements.txt
Click==7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
Pygments==2.5.2
Werkzeug==0.16.0
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
black==19.10b0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
dill==0.3.1.1
gast==0.2.2
google-auth-oauthlib==0.4.1
google-auth==1.10.0
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
ipython-genutils==0.2.0
ipython==7.10.2
jedi==0.15.2
langua==1.0.12
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandarallel==1.4.2
pandas==0.25.3
parso==0.5.2
pathspec==0.6.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1-modules==0.2.7
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2019.3
regex==2019.12.20
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.0
six==1.13.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-text==2.0.1
termcolor==1.1.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
